requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Configuration & Validation
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Resume Processing
PyMuPDF==1.23.8
python-docx==1.1.0
python-magic-bin==0.4.14
pdfplumber==0.10.3

# ML/NLP for Matching
sentence-transformers==2.2.2
torch==2.1.1
scikit-learn==1.3.2
spacy==3.7.2

# Additional NLP
transformers==4.35.2
nltk==3.8.1

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Caching & Background Jobs
redis==5.0.1
celery==5.3.4

# Vector Database (for advanced features)
chromadb==0.4.18

# Reporting & Visualization
openpyxl==3.1.2
reportlab==4.0.7
matplotlib==3.8.2
plotly==5.18.0

# Email & Templates
email-validator==2.1.0
jinja2==3.1.2

# Additional ML (Phase 1B Explainability)
lime==0.2.0.1
shap==0.44.0

# API & CORS
python-multipart==0.0.6
aiofiles==23.2.1

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
flake8==6.1.0

# Monitoring (Optional)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
